refactor: move LogicGameFlow patches to new file
